requirements.txt
certifi==2019.6.16
dj-database-url==0.5.0
Django==2.2.3
djangorestframework==3.10.1
gunicorn==19.9.0
Pygments==2.4.2
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.3
wincertstore==0.2
